Build WGPU native from a submodule
Added WGPU v24.0.0.2 directly as a submodule and build it as a library
(TODO: only for linux for now).

GLFW was updated to v3.4.

A copy of glfw3webgpu was taken that makes use of the new webgpu.h
header and glfw v3.4. The glfw3webgpu submodule was removed.

Emscripten does not support the latest webgpu.h header yet so a define
EMSCRIPTEN_WEBGPU_DEPRECATED was created for easy updating when the
header file does get updated.

Use of WGPUFuture does compile with WGPU native but an exception is
raised when trying to use wgpuWaitAny() stating that it is
unimplemented. A define for this has been created as well as
WEBGPU_FUTURE_UNIMPLEMENTED.