Send multiple uniforms in a struct Uniforms, including gamma, ratio and color
Gamma value depends on the texture format of the surface. Output from
fragment shader is expected to be linear, so if texture format is sRGB
then we need to perform gamma correction.

As per the WGSL spec, vec4f need to be aligned to 16 byte boundaries.
This affects the size of the structs and both the WGSL and Cpp struct
must match in size and in alignment of all members.
See https://www.w3.org/TR/WGSL/#alignment-and-size and
https://www.w3.org/TR/WGSL/#address-space-layout-constraints